Add functionality to verify the runtime environment
This will ensure that all dependencies of ez-build are installed,
yet still allow for the overriding of dependencies in the target
environment. It works by checking the installed packages of the
project (not necessarily what's in package.json) and verifies the
information against a hard coded list of dependencies.

(N.B.: we may want to switch this hardcoded list to just using
npm's peerDependencies feature. This will also not automatically
install the peer dependencies, but we make it more explicit what
dependencies are in fact involved.)

If all dependencies are installed, nothing happens; but for any
dependency that is missing ez-build will automatically call npm
to install it before building anything. On subsequent runs, no
installation will happen as long as the dependency is not removed.

This will not modify the project's package.json.

Ideally, this will mean projects do not directly depend on babel
and plugins thereof, but rather implicitly depend on them.
However, should projects depend on *additional* plugins, such as
the react preset, they must still include these in their list of
dev dependencies.