Pipeline redesign
This is a significant but mostly backwards compatible rewrite of
ez-build. We do away with the `--presets` and `--plugins` options
as described in #2, in favor of using tool-dependant RC files.
Another option is to mimic the interface of gcc, clang, and others,
whereby there is a specific CLI option to pass through parameters.
It could look something like thus:

```bash
$ ez-build -X js:option=value -X css:option=value
```

This continues the newly introduced interface of namespaced options,
i.e. `js:` or `css:`. This is now how `--include` and `--exclude`
works, and if the namespace isn't provided (as in the default
exclude value) it is assumed the user wants the value in all of
the namespaces. (Currently, only js and css, but we may well
introduce other pipelines, such as image optimizations etc.)

This rework also significantly improves our chances of performant
success in interactive builds (which this commit breaks, btw) since
we now pass single inputs to the compiler, and run things in parallel.
The exception to the parallel run is file writes, which happens
syncronously because FUD. Interactive builds will, from this point
on essentially mean ez-build takes care of the file watching itself
and then passes whatever files changed into the various pipelines.
This means we can do things like throttle builds to not happen
on fast changes, or otherwise be smarter about when things are
done.

We also introduce JSON logging with this commit, which is intended
as a tool for composing ez-build with other things. For instance,
ez-build currently outputs lines to say what input caused output.
A development server may choose to use this information to push
new versions to the client, enabling hot reloading of code and
content.

With this commit the following things are broken and must be
fixed before merging to master:

- Interactive builds
- Optimized builds