Refactor css url rebasing
The postcss-assets-rebase plugin didn't seem to work correctly in
all cases, and also seemed to use out-of-date APIs of postcss.
This change replaces that plugin with postcss-url, and implements
the resolution logic as a function.

We also make sure to use `opts.lib` so we don't hardcode paths,
in case they've been changed by the user.