When setting lib to project root, don't introduce extra slashes
If you specify `-L <project root>` when running ez-build, you'll
run into the situation of module names including an extra slash.
To fix this we simply check if `opts.lib` is set (it'll always
be relative to project root, so it'll be empty) and if not we
only use the project name as the module root.