Exclude .{git,npm}ignore contents and trim optimised-modules.json
The contents of .gitignore and .npmignore tends to be non-source
files, so we automatically exclude those. To re-enable, one will
have to add those files to the include patterns.

Additionally, we optimize optimised-modules.json to only include
those files that ez-build actually outputs, so any other files
should be ignored.