Implement option for backwards compatible default exports
This adds a precedence for flags that have a very granular purpose,
and that may or may not continue to work over time. Right now, the
only flag we care about is `add-module-exports` which enables us to
keep backwards compatibility with AMD and CJS consumers, who won't
know about the `{ default: export }` convention.