Separate code coverage from normal testing
This reverts the change to make unit tests load JIT compiled code
and instead loads AOT compiled code instead. This makes sense, as
we want to test the output, not precompiled source, and make sure
we never get out of sync in terms of configuration.

As well, this makes code coverage a distinct step from normal
testing, so as not to incur performance penalties in an iterative
development process.