Use `babel-preset-latest` in ez-build JS pipeline
This new preset affords us a few good things:

- Nice opt-in to es2017 via `--flags es2017:true`
- Easier management of ratified ES specs by only needing one preset
- An easy way to switch module format via `--flags modules:<format>`

The latter point enables us to more easily use ez-build with tooling
that expects a different module format, such as piping to Rollup.

Note: .babelrc will only be picked up when NODE_ENV is set to
anything other than `test`, which is necessary to ensure we don't
pick up ez-build's .babelrc when running tests. We may want to
reconsider this.

This fixes #34.