Add feature to read CLI options from files
Inspired by this [comment][1] on #27, this change adds the ability
for ez-build to read options from files, and placing their contents
precisely in the order the user intended. This is essentially the
same as just printing the contents of the files verbatim, taking
care to allow newlines by splitting on *all* whitespace.

This allows us to re-use ez-build configuration across scripts,
in a very predictable and explicit manner, as opposed to having
magical lookup rules that can be hard to reason about.

[1]: https://github.com/zambezi/ez-build/issues/27#issuecomment-247125657