Add simple React test case
This can probably be expanded significantly, but a decent start.
The tests break from other tests in this suite, in that it relies
on the fixture's build script for settings. This has the benefit
of not needing to update the tests because we change the fixture,
but on the other hand we have to add some logic to make sure we
link the accurate ez-build version when testing.

We also need to ensure we install npm dependencies, but this adds
considerable overhead to the execution time of the test,
so we only do it in case `node_modules/` doesn't already exist.