Conditionally load .babelrc, and display deprecation notice
Because removing .babelrc is a significant breaking change, we
add conditional loading of .babelrc files and print a big colorful
deprecation notice whenever a .babelrc (or package.babel key) is
detected.

This will give users some time to migrate their setups, and afford
us some time to enable functionality that may otherwise be missing
but should be included.