Resolve relative module source to fully qualified
The global identifier used to retrieve module dependencies is broken,
so this is an attempt to fix that by making sure we always fully
qualify the imports. For an author, it is transparent, they still use
relative imports as usual but the compilation process resolves them.

This fixes the global identifiers, but unfortunately break commonjs
require semantics, save from a specific node setup.

We will fork the transform-es2015-modules-umd plugin to fix the
underlying issue.