Fix real nasty file write issue
When using `output-file`, there's a chance creating the directory
for the file to be written fails, but this doesn't propagate up
the stack. This failure is silently ignored and a subsequent
write of the file fails, also silently. For some reason,
this never throws an error so the process exits with a zero
exit code. Only found out through the use of dtruss, that the
underlying `mkdir()` syscall returned -1, but I don't know why
since the error code doesn't propagate.

So we replace `output-file` with `mkdirp` and `write-file-atomic`.
This seems to fix the reason why the `mkdir()` syscall fails,
so I'm inclined to think the issue is in `output-file`, but I don't
actually know *why* it fails in the first place. Hopefully by
separating the dir creation and file write, we might get actual
errors propagated should this rear its ugly head again.