Refactor tests to make them simpler
This introduces a few more assert helpers, the most interesting
perhaps being the `assert_expected <actual>` helper, which when
first run will generate a file with the output of `<actual>` but
on subsequent runs will compare `<actual>` with the previously
stored data. If the data needs updating, just remove the relevant
files.

The `assert_expected` helper can be called any number of times
in a single test.