Improve flags and react documentation, and remove babelrc info
The babelrc docs were largely correct still, but only until such time that
v1.0.0 comes around, after which we'll stop reading babelrc files.
Because of this I just removed the docs, leaving any additional info to the
warnings printed to stdout when relying in .babelrc files.

The flags documentation were expanded to include all currently supported
flags, but removing the `es2017` information since it's now a ratified
standard and should be enabled by default.

The documentation on React support was incidental, and mostly covered using
babelrc files. Since this was removed, I rewrote that section to discuss
React and JSX.