Fix failing interactive exec test
Due to a change of the typical-project fixture, a test of the interactive
exec feature started failing. Arguably, this test was bad to begin with,
relying on the frailty of the typical-project fixture to simulate a failing
build. Instead, we simply change some of the code in the typical-project
fixture in the test, to provoka a build failure and thus create the context
for the interactive exec test.